feat(workspace): additive pnpm + Turborepo JS/TS monorepo membrane
Introduce a non-destructive monorepo membrane over the existing JS/TS islands
without touching the Python core, GOLD governance, or IMMUTABLE receipts.

Workspace:
- pnpm-workspace.yaml (ui-app, packages/*), pinned pnpm@10.33.0
- turbo.json (tasks API; globalDependencies on tsconfig.base.json;
  build/lint/typecheck/test), turbo pinned as root devDependency
- tsconfig.base.json: shared strict baseline

Shared packages:
- @enta/tsconfig: shared compiler covenant (base.json, nextjs.json)
- @enta/types: single source of truth for domain types (mirrors VOIDMAP.yml
  schema as a read-only consumer); type-only, private

ui-app wiring (ANNEX edits only):
- consumes @enta/types via @/types re-export (export type *)
- extends @enta/tsconfig/nextjs.json; adds typecheck script
- turbopack.root -> pnpm workspace root (fixes monorepo package resolution)
- noUncheckedIndexedAccess deferred for legacy ui-app code (roadmap)

Governance artifacts:
- SYNTHBIOSIS.md (architecture axioms, CEI), EPISTEMIC_HYGIENE.md
  (naming/claim-tagging/safe-extension), diagrams/monorepo-topology.md

Hygiene (G3): npm lockfiles moved to NICHTRAUM/archive/ with a reason note;
pnpm-lock.yaml is the single active lockfile.

CI (G6): .github/workflows/ci-js-workspace.yml runs turbo typecheck/lint/build
over the full workspace membrane; independent of the Python workflows.

Verified: pnpm install, turbo typecheck/lint/build all green; no .py, GOLD, or
receipt files modified.
