+++ packages/tsconfig/base.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["es2022"],
+    "target": "es2022",
+    "module": "esnext",
+    "moduleResolution": "bundler"
+  }
+}
