+++ packages/tsconfig/nextjs.json
@@ -1,0 +1,19 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "extends": "./base.json",
+  "compilerOptions": {
+    // ROADMAP: legacy Next app code predates noUncheckedIndexedAccess; it is
+    // relaxed here so the preset is non-breaking. New shared packages inherit
+    // the strict flag via base.json. Flip to true once ui-app is migrated.
+    "noUncheckedIndexedAccess": false,
+    "lib": ["dom", "dom.iterable", "esnext"],
+    "target": "ES2017",
+    "allowJs": true,
+    "noEmit": true,
+    "module": "esnext",
+    "moduleResolution": "bundler",
+    "jsx": "react-jsx",
+    "incremental": true,
+    "plugins": [{ "name": "next" }]
+  }
+}
