+++ packages/tsconfig/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@enta/tsconfig",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Shared TypeScript compiler configs for the EntaENGELment JS/TS workspace (internal ANNEX building block, never published).",
+  "exports": {
+    "./base.json": "./base.json",
+    "./nextjs.json": "./nextjs.json"
+  },
+  "files": ["base.json", "nextjs.json"]
+}
